backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
pydantic>=2.7.4
langchain>=0.2.5
langchain-groq>=0.1.6
langchain-core>=0.2.9
langgraph>=0.1.14
python-dotenv>=1.0.1